+++ frontend/webpack.dev.js
@@ -4,6 +4,9 @@ const common = require('./webpack.common');
 module.exports = merge(common, {
   mode: 'development',
   devtool: 'inline-source-map',
+    watchOptions: {
+      ignored: /node_modules/,
+    },
   devServer: {
     port: 8022,
     allowedHosts: ['localhost'],